Datasets use atomic write when persisting to disk
Partially fixes: #1954

Prior to this change dataset.download() would use a normal file write to
persist downloaded data to disk.

This meant another process or thread could check for the file and
attempt to read it before the full content was written.

This change uses a temporary file + a rename to update the file
atomically.

If a process is already reading a file that the new verion overwrites,
the previous file node is unlinked rather than being overwritten so the
read will work as expected.

This will allow us to back out optimistically pre-loading dataset data
before it is needed (and causing 404 errors when running tests on
machines without the appropriate permissions to download UK data)